+++ tests/unit/__main__.py
@@ -15,7 +15,9 @@ suite = unittest.TestLoader().loadTestsFromNames(
      'tests.unit.test_access.suite',
      'tests.unit.test_config.suite',
      'tests.unit.test_mib_controller.suite',
-     'tests.unit.test_cmd_responder.suite',
+     'tests.unit.test_snmp_responder.suite',
+     'tests.unit.test_snmp_notificator.suite',
+     'tests.unit.test_rest_responder.suite',
      'mapping_modules',
      'commands']
 )
